+++ SetupConfig.py
@@ -61,6 +61,7 @@ classifiers = [line.strip() for line in '''\
   Programming Language :: Python :: 3.10
   Programming Language :: Python :: 3.11
   Programming Language :: Python :: 3.12
+  Programming Language :: Python :: 3.13
   Topic :: Internet :: WWW/HTTP
   Topic :: Internet :: WWW/HTTP :: Dynamic Content
   Topic :: Internet :: WWW/HTTP :: Site Management