+++ docs/news.rst
@@ -16,8 +16,16 @@ Bug fixes:
     in ``PyMapping_HasKeyString`` under Python 3.13+
     caused by ``_namemapper`` looking up a key in a non-dictionary.
 
+  - Fixed ``_namemapper.c``: Silence ``IndexError`` when testing
+    ``name[attr]``. Some objects like ``re.MatchObject`` implement both
+    attribute access and index access. This confuses ``NameMapper`` because
+    it expects ``name[attr]`` to raise ``TypeError`` for objects that don't
+    implement mapping protocol.
+
   - Fixed mapping test in ``NameMapper.py``:
     Python 3.13 brough a new mapping type ``FrameLocalsProxy``.
+
+  - Fixed another ``RecursionError`` in ``ImportHooks`` under PyPy3.
 
 Tests:
 
@@ -25,8 +33,9 @@ Tests:
 
 CI:
 
-  - GHActions: Temporary run tests on macos-12. See
-    https://github.com/s-weigand/setup-conda/issues/432 for the reason.
+  - CI(GHActions): Switch to ``setup-miniconda``.
+
+  - CI(GHActions): Run tests under Python 3.13.
 
 3.3.3.post1 (2024-02-28)
 ------------------------